Added the game file to repo
The zip file contains the game.
About the game:
Basically a survival game where one moves with W A S D, shift for sprint, lctrl for sneak and basically a roguelike game, hence the name "A rogue with hunger". The game is free to play and was made using GMS2.